+++ sdk/integration-tests/package.json
@@ -10,13 +10,13 @@
     "wagmi": "^2.14.15"
   },
   "devDependencies": {
+    "@omni-network/test-utils": "workspace:*",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/react": "^16.2.0",
-    "@types/node": "^22.13.10",
+    "@types/react": "^19.0.8",
     "@vitest/coverage-v8": "^3.0.7",
     "happy-dom": "^17.4.4",
     "typescript": "^5.7.2",
-    "viem": "^2.23.13",
     "vite": "^6.2.0",
     "vitest": "^3.0.7"
   }
