+++ sdk/integration-tests/suites/flows.tsx
@@ -1,9 +1,5 @@
 import { type Quote, useQuote, useValidateOrder } from '@omni-network/react'
-import { act, waitFor } from '@testing-library/react'
-import { describe, expect, test } from 'vitest'
-
 import {
-  type AnyOrder,
   ETHER,
   INVALID_TOKEN_ADDRESS,
   MOCK_L1_ID,
@@ -11,10 +7,16 @@ import {
   OMNI_DEVNET_ID,
   TOKEN_ADDRESS,
   ZERO_ADDRESS,
-  createRenderHook,
-  executeTestOrder,
   mintOMNI,
   testAccount,
+} from '@omni-network/test-utils'
+import { act, waitFor } from '@testing-library/react'
+import { describe, expect, test } from 'vitest'
+
+import {
+  type AnyOrder,
+  createRenderHook,
+  executeTestOrder,
   testConnector,
   useOrderRef,
 } from '../test-utils.js'
