+++ sdk/integration-tests/suites/hooks.tsx
@@ -1,10 +1,5 @@
 import { useQuote } from '@omni-network/react'
-import { renderHook, waitFor } from '@testing-library/react'
-import { describe, expect, test } from 'vitest'
-
 import {
-  type AnyOrder,
-  ContextProvider,
   ETHER,
   INVALID_CHAIN_ID,
   INVALID_TOKEN_ADDRESS,
@@ -12,8 +7,15 @@ import {
   MOCK_L2_ID,
   TOKEN_ADDRESS,
   ZERO_ADDRESS,
-  executeTestOrder,
   testAccount,
+} from '@omni-network/test-utils'
+import { renderHook, waitFor } from '@testing-library/react'
+import { describe, expect, test } from 'vitest'
+
+import {
+  type AnyOrder,
+  ContextProvider,
+  executeTestOrder,
 } from '../test-utils.js'
 
 describe('useQuote()', () => {
